Various cleanups to ensure wc/settings.h and wc/types.h are in order.
