+++ wolfhsm/wh_server_she.h
@@ -25,9 +25,11 @@
 
 #include <stdint.h>
 
+/* Get server config */
 #include "wolfhsm/wh_server.h"
 
 #ifndef WOLFHSM_CFG_NO_CRYPTO
+#include "wolfssl/wolfcrypt/settings.h"
 #include "wolfssl/wolfcrypt/types.h"
 #include "wolfssl/wolfcrypt/aes.h"
 #include "wolfssl/wolfcrypt/cmac.h"
